Convert ByteArray to Image before generating PNG
This extra layer of abstraction allows us to generate different Image
structures (i.e. different ways of representing the bytes), without changing the
Image to PNG code.

Since the Image contains the color data and dimensions, we could also use it to
display it in other places, like an interactive SDL window (and still allowing
PNG exports with the image2png function).